Count Cleared Rows #13
Add support for a seven-segment display that shows a count of the rows
that are completed and
cleared. (This functionality requires that the clearing of completed
rows feature has been
implemented.) The count should be a two digit number (0 to 99) with both
digits displayed
(although you may choose not to display the left digit for counts in the
range 0 to 9 if you wish).
The count should start at 0 and be incremented by 1 every time a row is
cleared. The count
should be capped at 99. It must be possible to play your game so that at
least 10 rows are cleared
(i.e. both digits in this count change). The display must be maintained
when the game is paused.
Your feature summary form must indicate which AVR pins the seven-segment
display pins are
connected to.